Feature/item tipo crud

+++ src/main/java/com/controle_de_gastos/notas_api/service/ItemTipoService.java
@@ -1,11 +1,17 @@
 package com.controle_de_gastos.notas_api.service;
 
+import com.controle_de_gastos.notas_api.dto.projecao.ItemSimplesProjecaoDTO;
 import com.controle_de_gastos.notas_api.dto.requisicao.ItemTipoRequisicaoDTO;
+import com.controle_de_gastos.notas_api.dto.resposta.ItemTipoComItemRespostaDTO;
+import com.controle_de_gastos.notas_api.model.Item;
 import com.controle_de_gastos.notas_api.repository.ItemTipoRepository;
 import com.controle_de_gastos.notas_api.dto.resposta.ItemTipoRespostaDTO;
 import com.controle_de_gastos.notas_api.model.ItemTipo;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -35,13 +41,73 @@ public class ItemTipoService {
     }
 
     public ItemTipoRespostaDTO criar(ItemTipoRequisicaoDTO tipoDTO){
-        ItemTipo tipo = new ItemTipo();
-        tipo.setNome(tipoDTO.nome());
+        ItemTipo tipo = ItemTipo.builder()
+                .nome(tipoDTO.nome())
+                .build();
        return toRespostaDTO(itemTipoRepository.save(tipo));
     }
 
-    public void deletarPorId(Integer id){
+    public Optional<ItemTipoRespostaDTO> atualizarTudo(Integer id, ItemTipoRequisicaoDTO tipoDTO){
+        return itemTipoRepository.findById(id)
+                .map(tipo -> {
+                    tipo.setNome(tipoDTO.nome());
+                    return toRespostaDTO(itemTipoRepository.save(tipo));
+                });
+    }
+
+    public List<ItemTipoComItemRespostaDTO> listarItensPorItemTipoTodos(){
+        return itemTipoRepository.findAll()
+                .stream()
+                .map(tipo ->{
+                    return new ItemTipoComItemRespostaDTO(
+                       tipo.getId(),
+                       tipo.getNome(),
+                       tipo.getItens()
+                               .stream()
+                               .map(item->{
+                                   return new ItemSimplesProjecaoDTO(
+                                           item.getId(),
+                                           item.getNome(),
+                                           item.getPeso(),
+                                           item.getVersao()
+                                   );
+                                       }).toList()
+               );
+                })
+                .toList();
+
+    }
+
+    public Optional<ItemTipoComItemRespostaDTO> listarItensPorItemTipoId(Integer id){
+        return itemTipoRepository.findById(id)
+                .map(tipo -> {
+                   return new ItemTipoComItemRespostaDTO(
+                            tipo.getId(),
+                            tipo.getNome(),
+                            tipo.getItens()
+                                    .stream()
+                                    .map(item ->{
+                                        return new ItemSimplesProjecaoDTO(
+                                                item.getId(),
+                                                item.getNome(),
+                                                item.getPeso(),
+                                                item.getVersao()
+                                        );
+                                    })
+                                    .toList()
+                    );
+                });
+    }
+
+
+    public boolean deletarPorId(Integer id){
+        Optional<ItemTipo> itemTipo = itemTipoRepository.findById(id);
+        if(itemTipo.isEmpty()) return false;
+
+        if (!itemTipo.get().getItens().isEmpty()) throw new IllegalStateException("Não é possível deletar, existem dependências");
+
         itemTipoRepository.deleteById(id);
+        return true;
     }
 
 }
